Test that the cart can be created when logged out
We expect to be able to create a cart without having a user. Document
this by adding a test.

Due to interactions with the `rspec_api_documentation` gem, refactor the
test to inline the memoized methods. This might increase readability,
too.

Clarify that the `user_id` parameter is optional by specifying that the
`origin` parameter is required.

https://app.clubhouse.io/bark/story/10595/fe-update-subscription-checkout-to-use-existing-cart-endpoints